Extract shared private helpers into per-helper partial files
Per Jake's review feedback, private helpers used by multiple operators get their
own ObservableCacheEx.{HelperName}.cs file, matching the per-operator pattern
established for the public surface.

  Combine                    -> ObservableCacheEx.Combine.cs (from And.cs)
  AdaptSelector              -> ObservableCacheEx.AdaptSelector.cs (from Group.cs)
  OnChangeAction             -> ObservableCacheEx.OnChangeAction.cs (from OnItemAdded.cs)
  ForForced                  -> ObservableCacheEx.ForForced.cs (from Transform.cs)
  CreateChangeSetTransformer -> ObservableCacheEx.CreateChangeSetTransformer.cs (from TransformManyAsync.cs)
  TrueFor                    -> ObservableCacheEx.TrueFor.cs (from TrueForAll.cs)

DefaultSortResetThreshold const moved to ObservableCacheEx.cs (the core file).
Audit found it is used by both Sort and SortBy, contrary to the original PR body.

AsyncDisposeMany #if SUPPORTS_ASYNC_DISPOSABLE wrapping replaced with a project-level
Compile Remove. The file body is unconditionally compiled on supported platforms and
excluded entirely on unsupported ones (net4*).

All extractions are byte-preserving moves with no functional change. Builds clean on
all target frameworks (netstandard2.0, net462, net6-net10). Targeted tests pass.
