+++ src/DynamicData/Cache/ObservableCacheEx.AdaptSelector.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2011-2025 Roland Pheasant. All rights reserved.
+// Copyright (c) 2011-2025 Roland Pheasant. All rights reserved.
 // Roland Pheasant licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
@@ -25,5 +25,9 @@ namespace DynamicData;
 /// </summary>
 public static partial class ObservableCacheEx
 {
-    private const int DefaultSortResetThreshold = 100;
+    // TODO: Apply the Adapter to more places
+    private static Func<TObject, TKey, TResult> AdaptSelector<TObject, TKey, TResult>(Func<TObject, TResult> other)
+        where TObject : notnull
+        where TKey : notnull
+        where TResult : notnull => (obj, _) => other(obj);
 }
