+++ src/DynamicData/Cache/ObservableCacheEx.TrueForAll.cs
@@ -75,9 +75,4 @@ public static partial class ObservableCacheEx
         where TObject : notnull
         where TKey : notnull
         where TValue : notnull => source.TrueFor(observableSelector, items => items.All(o => o.LatestValue.HasValue && equalityCondition(o.Item, o.LatestValue.Value)));
-
-    private static IObservable<bool> TrueFor<TObject, TKey, TValue>(this IObservable<IChangeSet<TObject, TKey>> source, Func<TObject, IObservable<TValue>> observableSelector, Func<IEnumerable<ObservableWithValue<TObject, TValue>>, bool> collectionMatcher)
-        where TObject : notnull
-        where TKey : notnull
-        where TValue : notnull => new TrueFor<TObject, TKey, TValue>(source, observableSelector, collectionMatcher).Run();
 }
